src/requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-SocketIO==4.2.1
idna==2.8
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
netifaces==0.10.6
nose==1.3.7
numpy==1.18.1
opencv-python==4.2.0.32
Pillow==7.0.0
python-engineio==3.11.2
python-socketio==4.4.0
requests==2.22.0
six==1.14.0
socketio==0.2.1
tornado==6.0.2
urllib3==1.25.6
websockets==7.0
Werkzeug==0.16.0
